+++ frontend/src/pages/today-focus/TodayFocusPoint.module.css
@@ -1,0 +1,47 @@
+.pointContainer {
+  display: flex;
+  flex-direction: column;
+  row-gap: 8px;
+  margin: 16px 0 24px 0;
+}
+
+.pointTxt {
+  font-weight: 400;
+  font-size: 18px;
+  color: var(--text-gray);
+}
+
+.point {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  border: 1px, solid, var(--color-gray-200);
+  border-radius: 50px;
+  padding: 8px 12px;
+  gap: 4px;
+  height: 35px;
+  margin-right: auto;
+}
+
+.point img {
+  width: 19px;
+  height: 19px;
+}
+
+.pointTotalTxt {
+  font-weight: 500;
+  font-size: 16px;
+  color: var(--text-black);
+}
+
+@media (min-width: 744px) {
+  .pointContainer {
+    margin: 16px 0 40px 0;
+  }
+}
+
+@media (min-width: 1200px) {
+  .pointContainer {
+    margin: 16px 0 20px 0;
+  }
+}
