+++ frontend/src/pages/today-focus/TodayFocusTimer.module.css
@@ -1,0 +1,108 @@
+.focusSelectTimeContainer {
+  position: absolute;
+  top: 45px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px, solid, var(--color-gray-200);
+  border-radius: 50px;
+  padding: 5px 8px;
+  gap: 4px;
+  visibility: hidden;
+  opacity: 0;
+  transition: opacity 0.3s, visibility 0s 0.3s;
+}
+
+.focusSelectTimeContainer p {
+  font-weight: 500;
+  font-size: 16px;
+}
+
+.focusSelectTimeContainer.show {
+  visibility: visible;
+  opacity: 1;
+  transition: opacity 0.5s, visibility 0s;
+}
+
+.focusTimerContainer {
+  display: flex;
+  justify-content: center;
+  width: 280px;
+  color: var(--text-black);
+  font-weight: 800;
+  font-size: 80px;
+}
+
+.focusMinTimer {
+  background-color: transparent;
+  text-align: end;
+  color: var(--text-black);
+  font-weight: 800;
+  font-size: 80px;
+  width: 100%;
+  border: none;
+  outline: none;
+  padding: 0;
+}
+
+.focusSecTimer {
+  background-color: transparent;
+  text-align: start;
+  color: var(--text-black);
+  font-weight: 800;
+  font-size: 80px;
+  width: 100%;
+  border: none;
+  outline: none;
+  padding: 0;
+}
+
+.countDown {
+  color: var(--error-light);
+}
+
+.countDown input {
+  color: var(--error-light);
+}
+
+.timeover {
+  color: var(--text-gray);
+}
+
+.timeover input {
+  color: var(--text-gray);
+}
+
+@media (min-width: 744px) {
+  .focusSelectTimeContainer {
+    top: 53px;
+  }
+
+  .focusTimerContainer {
+    font-size: 120px;
+    width: 430px;
+  }
+
+  .focusMinTimer {
+    font-size: 120px;
+  }
+
+  .focusSecTimer {
+    font-size: 120px;
+  }
+}
+
+@media (min-width: 1200px) {
+  .focusTimerContainer {
+    font-size: 150px;
+    width: 517px;
+  }
+
+  .focusMinTimer {
+    font-size: 150px;
+  }
+
+  .focusSecTimer {
+    font-size: 150px;
+  }
+}
