+++ frontend/src/pages/today-focus/TodayFocusTimerBtn.module.css
@@ -1,0 +1,141 @@
+.focusBtnContainer {
+  display: flex;
+  gap: 8px;
+  justify-content: center;
+  align-items: center;
+}
+
+.focusPauseBtn {
+  background-color: var(--text-green);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: none;
+  border-radius: 100%;
+  padding: 0;
+  width: 48px;
+  height: 48px;
+  cursor: pointer;
+  visibility: hidden;
+  opacity: 0;
+  transition: opacity 400ms, visibility 0s 400ms;
+}
+
+.focusPauseBtn img {
+  width: 26.4px;
+  height: 26.4px;
+}
+
+.focusPauseBtn.show {
+  visibility: visible;
+  opacity: 1;
+  transition: opacity 0.5s, visibility 0s;
+}
+
+.focusStartBtn {
+  color: var(--text-white);
+  background-color: var(--bg-green);
+  font-weight: 800;
+  font-size: 18px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: none;
+  border-radius: 50px;
+  padding: 10px 28px 10px 20px;
+  gap: 8px;
+  cursor: pointer;
+  height: 45px;
+  margin-top: 1px;
+  margin-bottom: 2px;
+}
+
+.focusStartBtn:hover {
+  background-color: var(--text-green);
+  transition: all, 500ms;
+}
+
+.focusStartBtn.disable {
+  background-color: var(--text-gray);
+  cursor: default;
+}
+
+.focusStartBtn p {
+  width: 60px;
+}
+
+.focusStartBtn img {
+  width: 24px;
+  height: 24px;
+}
+
+.focusResetBtn {
+  background-color: var(--bg-green);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: none;
+  border-radius: 100%;
+  padding: 0;
+  width: 48px;
+  height: 48px;
+  cursor: pointer;
+  visibility: hidden;
+  opacity: 0;
+  transition: opacity 400ms, visibility 0s 400ms;
+}
+
+.focusResetBtn img {
+  width: 26.4px;
+  height: 26.4px;
+}
+
+.focusResetBtn.show {
+  visibility: visible;
+  opacity: 1;
+  transition: opacity 0.5s, visibility 0s;
+}
+
+@media (min-width: 744px) {
+  .focusBtnContainer {
+    gap: 24px;
+  }
+
+  .focusPauseBtn {
+    width: 64px;
+    height: 64px;
+  }
+
+  .focusPauseBtn img {
+    width: 35.2px;
+    height: 35.2px;
+  }
+
+  .focusStartBtn {
+    font-size: 28px;
+    padding: 8px 105px 8px 78px;
+    gap: 16px;
+    height: 60px;
+    margin-top: 0px;
+    margin-bottom: 4px;
+  }
+
+  .focusStartBtn p {
+    width: 90px;
+  }
+
+  .focusStartBtn img {
+    width: 44px;
+    height: 44px;
+  }
+
+  .focusResetBtn {
+    width: 64px;
+    height: 64px;
+  }
+
+  .focusResetBtn img {
+    width: 35.2px;
+    height: 35.2px;
+  }
+}
